backend/requirements.txt
fastapi==0.123.5
uvicorn==0.38.0
python-multipart==0.0.20
openai-whisper==20250625
google-generativeai==0.8.5
numpy==2.0.2
SQLAlchemy==2.0.44
pyannote.audio==3.4.0
torch==2.5.1
torchaudio==2.5.1
supabase==2.25.0
python-dotenv==1.2.1
pydantic==2.12.5